backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
transformers==4.37.2
torch==2.2.0
torchaudio==2.2.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.3
python-dotenv==1.0.0
aiofiles==25.1.0
pyannote.audio==3.1.1